pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "plotly>=5.24.1",
    "psycopg2-binary>=2.9.10",
    "pywebpush>=2.0.1",
    "streamlit>=1.39.0",
    "websockets>=12.0",
    "asyncio>=3.4.3",
    "cssutils>=2.7.1",
    "jsbeautifier>=1.14.9",
    "pytz>=2023.3",
    "python-dateutil>=2.8.2"
]

[nix]
channel = "stable-24_05"
